+++ src/cutty/projects/common.py
@@ -1,0 +1,3 @@
+"""Common functionality for projects."""
+LATEST_BRANCH = "cutty/latest"
+UPDATE_BRANCH = "cutty/update"
