+++ tests/unit/projects/test_link.py
@@ -6,9 +6,9 @@ import pytest
 
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path as VirtualPath
+from cutty.projects.common import LATEST_BRANCH
+from cutty.projects.common import UPDATE_BRANCH
 from cutty.projects.create import CreateProject
-from cutty.projects.create import LATEST_BRANCH
-from cutty.projects.create import UPDATE_BRANCH
 from cutty.projects.link import linkproject
 from cutty.repositories.domain.repository import Repository as Template
 from cutty.util.git import Repository
